asdf

+++ HitMind with C.ver_2/HitMindwithCver2.c
@@ -249,7 +249,7 @@ int main(int argc, char *argv[])
 				}
 				//	}
 
-				
+
 				if (status.ishappen == true) {
 					if (sum) {
 						//i = Get_Notice_sql(cons, notice);
@@ -1442,7 +1442,7 @@ int main(int argc, char *argv[])
 			for (int i = 0; i < maxchating; i++) {
 				chatingH += HeightOfText(chatings[i].name, Display_X*0.65, renderer, chatings[i].message, 30 * ((float)Display_X / 1920), 1);
 			}
-			if (chatingH >chatlimith) {
+			if (chatingH > chatlimith) {
 				chatslide->End = chatingH - chatlimith;
 				MoveSlider_value(chatslide, chatslide->End);
 			}
@@ -1465,7 +1465,7 @@ int main(int argc, char *argv[])
 					for (int i = 0; i < maxchating; i++) {
 						chatingH += HeightOfText(chatings[i].name, Display_X*0.65, renderer, chatings[i].message, 30 * ((float)Display_X / 1920), 1);
 					}
-					if (chatingH >chatlimith) {
+					if (chatingH > chatlimith) {
 						chatslide->End = chatingH - chatlimith;
 						MoveSlider_value(chatslide, *chatslide->Value);
 						chatmovehappen = true;
@@ -1805,7 +1805,7 @@ int main(int argc, char *argv[])
 						for (int i = 0; i < maxchating; i++) {
 							chatingH += HeightOfText(chatings[i].name, Display_X*0.65, renderer, chatings[i].message, 30 * ((float)Display_X / 1920), 1);
 						}
-						if (chatingH >chatlimith) {
+						if (chatingH > chatlimith) {
 							chatslide->End = chatingH - chatlimith;
 							MoveSlider_value(chatslide, *chatslide->Value);
 							chatmovehappen = true;
@@ -3616,20 +3616,18 @@ int main(int argc, char *argv[])
 					}
 				}
 				PutText(renderer, "나가기", Display_X*0.807, Display_Y*0.75, 57 * ((float)Display_X) / 1920, 255, 255, 255, 1);
-				if (bangsang == 1 && isready)
+				if (bangsang == 1 && isready && gameuser[0].status != 1 && gameuser[1].status != 1 && gameuser[2].status != 1 && gameuser[3].status != 1)
 				{
-					if (gameuser[0].status != 1 && gameuser[1].status != 1 && gameuser[2].status != 1 && gameuser[3].status != 1)
+					if (PutRoundButton(renderer, 255, 0, 0, 210, 0, 0, 255, 0, 0, Display_X*0.7317, Display_Y*0.85, Display_X*0.2343, Display_Y*0.1157, 20, 0, &event, &happen)) //시작하기, 준비 버튼
 					{
-						if (PutRoundButton(renderer, 255, 0, 0, 210, 0, 0, 255, 0, 0, Display_X*0.7317, Display_Y*0.85, Display_X*0.2343, Display_Y*0.1157, 20, 0, &event, &happen)) //시작하기, 준비 버튼
-						{
-							send(ClientParam.Cconnect_socket, "game start", 40, 0);
-							sprintf(query, "delete from room where num = %d", My_Room.ownnum);
-							mysql_query(cons, query);
-							//		MouseUP_Wait;
-						}
-						PutText(renderer, "시작하기", Display_X*0.796, Display_Y*0.87, 57 * ((float)Display_X) / 1920, 255, 255, 255, 1);    //방장일때는 시작하기
-
+						send(ClientParam.Cconnect_socket, "game start", 40, 0);
+						sprintf(query, "delete from room where num = %d", My_Room.ownnum);
+						mysql_query(cons, query);
+						//		MouseUP_Wait;
 					}
+					PutText(renderer, "시작하기", Display_X*0.796, Display_Y*0.87, 57 * ((float)Display_X) / 1920, 255, 255, 255, 1);    //방장일때는 시작하기
+
+
 				}
 				else {
 					if (PutRoundButton(renderer, 255, 0, 0, 210, 0, 0, 255, 0, 0, Display_X*0.7317, Display_Y*0.85, Display_X*0.2343, Display_Y*0.1157, 20, 0, &event, &happen)) //시작하기, 준비 버튼
@@ -3846,6 +3844,16 @@ int main(int argc, char *argv[])
 					SDL_RenderPresent(renderer);
 					//printf("update %lld\n", ++ccount);
 					////printf("render	");
+					continue;
+				}
+				if (event.type == SDL_KEYDOWN) {
+					SDL_RenderReadPixels(renderer, &Displayrect, SDL_PIXELFORMAT_ARGB8888, SurfaceOfRenderer->pixels, SurfaceOfRenderer->pitch);
+					printf("saved\n");
+				}
+				if (event.type == SDL_WINDOWEVENT&& event.window.event == SDL_WINDOWEVENT_RESTORED &&Full) {
+					SDL_BlitSurface(SurfaceOfRenderer, NULL, SurfaceOfWindow, NULL);
+					SDL_UpdateWindowSurface(Window);
+					printf("restored\n");
 					continue;
 				}
 				if (event.type == SDL_MOUSEMOTION) {
@@ -4404,13 +4412,7 @@ int main(int argc, char *argv[])
 					break;
 				case SDL_WINDOWEVENT:
 					switch (event.window.event) {
-
-					case SDL_WINDOWEVENT_RESTORED: if (Full) {
-						SDL_BlitSurface(SurfaceOfRenderer, NULL, SurfaceOfWindow, NULL);
-						SDL_UpdateWindowSurface(Window);
 						break;
-					}
-												   break;
 					case SDL_WINDOWEVENT_CLOSE:// 다수 창에서의 닫기이벤트가 발생할경우
 						if (Me->Turn == 1) {// DB연동
 							int i = NowPlayer;
@@ -4726,6 +4728,11 @@ int main(int argc, char *argv[])
 				{
 					if (gameuser[i].status != 0)
 					{
+						if (gameuser[i].Th == 1 && gameuser[i].ownnum == myuser->ownnum)
+						{
+							sprintf(query, "update user set level = level+1 where ownnum = %d", myuser->ownnum);
+							mysql_query(cons, query);
+						}
 						FillRoundRect(renderer, 0, 176, 240, Display_X * 0.357, Display_Y * (0.37 + (0.09 * (i + 1))), Display_X * 0.286, Display_Y * 0.08, 18 * ((float)Display_X / 1920));
 						DrawRect(renderer, 0, 112, 192, Display_X * 0.452, Display_Y * (0.371 + (0.09 * (i + 1))), Display_X * 0.003, Display_Y * 0.079);
 						DrawRect(renderer, 0, 112, 192, Display_X * 0.545, Display_Y * (0.371 + (0.09 * (i + 1))), Display_X * 0.003, Display_Y * 0.079);
